requirements.txt
torch>=1.7.0
torchvision>=0.8.1
streamlit>=1.0.0
numpy>=1.18.0
matplotlib>=3.1.0
scikit-learn>=0.23.0
Pillow>=8.0.0
jupyter>=1.0.0
